requirements.txt
acres==0.5.0
aiofiles==24.1.0
altair==5.5.0
altgraph==0.17.4
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.0
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.2
cycler==0.12.1
etelemetry==0.3.1
filelock==3.18.0
fitz==0.0.1.dev2
fonttools==4.57.0
frontend==0.0.3
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
looseversion==1.3.0
lxml==6.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
narwhals==1.39.1
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.5
opencv-python==4.12.0.88
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pefile==2023.2.7
pillow==11.2.1
protobuf==6.31.0
prov==2.1.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
puremagic==1.30
pyarrow==20.0.0
pydeck==0.9.1
pydot==4.0.1
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.5
PyMuPDF==1.26.3
pyparsing==3.2.3
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
pyqtgraph==0.13.7
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-telegram-bot==22.2
pytz==2025.2
pywin32-ctypes==0.2.3
pyxnat==1.6.3
rdflib==7.1.4
referencing==0.36.2
reportlab==4.4.2
requests==2.32.3
rpds-py==0.25.0
scipy==1.16.0
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.47.1
streamlit==1.45.1
tabulate==0.9.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5
traits==7.0.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.35.0
watchdog==6.0.0
